Move reconcile algo to Portal Wrap updateScene with withBatchUpdates
